requirements.txt
flask
torch
pillow
transformers